Finished design of next section for all screen views.
I also added a scrollbar. I think a regular scroll is going to be better and faster to implement since we are pressed on time. I customized it so I think it still looks very clean.

I also changed the location of some files throughout the project, such as the server's node files.